Separate typography plugin tests
