+++ __fixtures__/pluginTypography/config.json
@@ -1,0 +1,3 @@
+{
+  "config": "__fixtures__/pluginTypography/tailwind.config.js"
+}
