+++ __fixtures__/pluginTypography/pluginTypography.js
@@ -1,0 +1,7 @@
+import tw from './macro'
+
+// From @tailwindcss/typography
+tw`prose sm:prose-sm lg:prose-lg xl:prose-xl`
+
+// From tailwindcss-typography
+tw`rich-text`
